feat: create add/mod.rs
